fix(snap): extract serveByteCodes to SnapServer; fix 2MB cap, 1024-hash cap, EmptyCodeHash return
go-ethereum ServiceGetByteCodesQuery (handlers.go:347-373) enforces three invariants that the
inline handleGetByteCodes implementation was missing:
  1. 2MB hard cap on responseBytes (Bug 1)
  2. maxCodeLookups = 1024 cap on request hash count (Bug 2)
  3. EmptyCodeHash returned as []byte{} without DB lookup (Bug 3)

Extraction to SnapServer.serveByteCodes brings handleGetByteCodes into line with the other three
SNAP handlers which already delegate to SnapServer.* methods.
